Reorganize optics instances per data type
