--- modules/optics/arrow-optics/src/main/kotlin/arrow/optics/instances/index_java_lang.kt
@@ -1,8 +1,0 @@
-package java_lang
-
-import arrow.optics.instances.StringIndexInstance
-
-object StringIndexInstanceImplicits {
-    @JvmStatic
-    fun instance(): StringIndexInstance = StringIndexInstance
-}
